Remove fast-tracking in TagMacro for Scala 3 to support custom Tags for abstract type members (#552)
* Remove fast-tracking in TagMacro for Scala 3 to support custom Tags for abstract type members

* invert condition
